+++ serde-encrypt-core/src/encrypt.rs
@@ -1,0 +1,5 @@
+//! Encryption/Decryption implementations.
+
+pub mod encrypted_message;
+pub mod plain_message_public_key;
+pub mod plain_message_shared_key;