+++ serde-encrypt-core/src/key.rs
@@ -1,0 +1,8 @@
+#![allow(missing_docs)]
+
+#[deny(missing_docs)]
+pub mod combined_key;
+#[deny(missing_docs)]
+pub mod key_pair;
+
+pub mod as_shared_key;