+++ src/key.rs
@@ -1,8 +1,0 @@
-#![allow(missing_docs)]
-
-#[deny(missing_docs)]
-pub mod combined_key;
-#[deny(missing_docs)]
-pub mod key_pair;
-#[deny(missing_docs)]
-pub mod shared_key;