+++ src/traits/impl_detail.rs
@@ -1,7 +1,0 @@
-pub(in crate::traits) mod nonce;
-
-mod serialize;
-
-pub use serialize::SerializedPlain;
-
-pub(in crate::traits) use serialize::serialize;